+++ pytest_assert_utils/util/decl.py
@@ -1,4 +1,5 @@
-import collections.abc
+import collections
+import inspect
 import typing
 
 __all__ = [
@@ -7,6 +8,8 @@ __all__ = [
     'Collection',
     'List',
     'Set',
+    'Dict',
+    'Str',
 ]
 
 
@@ -80,20 +83,23 @@ class Optional:
         return other in (None, self.value)
 
 
-class _CollectionValuesCheckerMeta(type(collections.abc.Collection)):
+class _CollectionValuesCheckerMeta(type(collections.Collection)):
     """Overloads isinstance() to enable truthiness based on values inside the collection
     """
+    def __eq__(cls: typing.Type['_CollectionValuesChecker'], instance) -> bool:
+        return cls.__instancecheck__(instance)
+
     def __instancecheck__(cls: typing.Type['_CollectionValuesChecker'], instance) -> bool:
-        if cls._must_be_empty and len(instance) > 0:
+        if cls._must_be_empty and not cls._collection_is_empty_(instance):
             return False
 
-        if cls._must_not_be_empty and len(instance) == 0:
+        if cls._must_not_be_empty and cls._collection_is_empty_(instance):
             return False
 
-        if cls._must_contain and not all(v in instance for v in cls._must_contain):
+        if cls._must_contain and not all(cls._collection_contains_(instance, v) for v in cls._must_contain):
             return False
 
-        if cls._must_not_contain and any(v in instance for v in cls._must_not_contain):
+        if cls._must_not_contain and any(cls._collection_contains_(instance, v) for v in cls._must_not_contain):
             return False
 
         return True
@@ -108,36 +114,42 @@ class _CollectionValuesCheckerMeta(type(collections.abc.Collection)):
             parts.append('not_empty()')
 
         if cls._must_contain:
-            items = [str(item) for item in cls._must_contain]
-            parts.append(f'containing({", ".join(items)})')
+            parts.append(f'containing({cls._repr_containing_(cls._must_contain)})')
 
         if cls._must_not_contain:
-            items = [str(item) for item in cls._must_not_contain]
-            parts.append(f'not_containing({", ".join(items)})')
+            parts.append(f'not_containing({cls._repr_not_containing_(cls._must_not_contain)})')
 
         return '.'.join(parts)
 
 
 class _CollectionValuesChecker:
-    _must_contain: typing.Set = frozenset()
-    _must_not_contain: typing.Set = frozenset()
+    _must_contain: typing.Tuple = ()
+    _must_not_contain: typing.Tuple = ()
     _must_be_empty: bool = False
     _must_not_be_empty: bool = False
 
     @classmethod
+    def _unpack_generators(cls, items):
+        for item in items:
+            if inspect.isgenerator(item):
+                yield from item
+            else:
+                yield item
+
+    @classmethod
     def containing(cls: T, *items) -> T:
-        _must_contain = {
+        _must_contain = (
             *cls._must_contain,
-            *items,
-        }
+            *cls._unpack_generators(items),
+        )
         return type(cls.__name__, (cls,), {'_must_contain': _must_contain})
 
     @classmethod
     def not_containing(cls: T, *items) -> T:
-        _must_not_contain = {
+        _must_not_contain = (
             *cls._must_not_contain,
-            *items,
-        }
+            *cls._unpack_generators(items),
+        )
         return type(cls.__name__, (cls,), {'_must_not_contain': _must_not_contain})
 
     @classmethod
@@ -148,77 +160,201 @@ class _CollectionValuesChecker:
     def not_empty(cls: T) -> T:
         return type(cls.__name__, (cls,), {'_must_not_be_empty': True})
 
+    @classmethod
+    def _repr_containing_(cls, must_contain):
+        return ', '.join(repr(item) for item in must_contain)
 
-class Collection(_CollectionValuesChecker, collections.abc.Collection, metaclass=_CollectionValuesCheckerMeta):
-    """Special class allowing utils.Any to check items in any collection (list, set, tuple, etc)
+    @classmethod
+    def _repr_not_containing_(cls, must_not_contain):
+        return cls._repr_containing_(must_not_contain)
+
+    @classmethod
+    def _collection_is_empty_(cls, collection) -> bool:
+        return len(collection) == 0
+
+    @classmethod
+    def _collection_contains_(cls, collection, v) -> bool:
+        return v in collection
+
+
+class _DictValuesChecker(_CollectionValuesChecker):
+    @classmethod
+    def _process_items(
+        cls,
+        *items: typing.Union[typing.Hashable, typing.Dict],
+    ) -> typing.List[typing.Tuple[typing.Hashable, typing.Any]]:
+        processed = []
+        for item in cls._unpack_generators(items):
+            if isinstance(item, dict):
+                processed.extend(item.items())
+            else:
+                processed.append((item, Any()))
+        return processed
+
+    @classmethod
+    def containing(cls: T, *items, **kwargs) -> T:
+        return super().containing(*cls._process_items(*items, kwargs))
+
+    @classmethod
+    def not_containing(cls: T, *items, **kwargs) -> T:
+        return super().not_containing(*cls._process_items(*items, kwargs))
+
+    @classmethod
+    def _repr_containing_(cls, must_contain):
+        return dict(must_contain)
+
+    @classmethod
+    def _collection_contains_(cls, collection, v) -> bool:
+        k, v = v
+        return k in collection and collection[k] == v
+
+
+class Collection(_CollectionValuesChecker, collections.Collection, metaclass=_CollectionValuesCheckerMeta):
+    """Special class enabling equality comparisons to check items in any collection (list, set, tuple, etc)
 
     Examples of functionality:
 
-    >>> Any(Collection.containing(1)) == [1, 2, 3]
+    >>> Collection.containing(1) == [1, 2, 3]
     True
-    >>> Any(Collection.containing(1)) == {1, 2, 3}
+    >>> Collection.containing(1) == {1, 2, 3}
     True
-    >>> Any(Collection.containing(1)) == (1, 2, 3)
+    >>> Collection.containing(1) == (1, 2, 3)
     True
 
-    >>> Any(Collection.containing(1)) == [4, 5, 6]
+    >>> Collection.containing(1) == [4, 5, 6]
     False
-    >>> Any(Collection.containing(1)) == {4, 5, 6}
+    >>> Collection.containing(1) == {4, 5, 6}
     False
-    >>> Any(Collection.containing(1)) == (4, 5, 6)
+    >>> Collection.containing(1) == (4, 5, 6)
     False
     """
 
 
 class List(_CollectionValuesChecker, list, metaclass=_CollectionValuesCheckerMeta):
-    """Special class allowing utils.Any to check items in a list
+    """Special class enabling equality comparisons to check items in a list
 
     Examples of functionality:
 
-    >>> Any(List.containing(1)) == [1, 2, 3]
+    >>> List.containing(1) == [1, 2, 3]
     True
-    >>> Any(List.containing(1)) == [4, 5, 6]
+    >>> List.containing(1) == [4, 5, 6]
     False
 
-    >>> Any(List.not_containing(1)) == [1, 2, 3]
+    >>> List.not_containing(1) == [1, 2, 3]
     False
-    >>> Any(List.not_containing(1)) == [4, 5, 6]
+    >>> List.not_containing(1) == [4, 5, 6]
     True
 
-    >>> Any(List.empty()) == [1, 2, 3]
+    >>> List.empty() == [1, 2, 3]
     False
-    >>> Any(List.empty()) == []
+    >>> List.empty() == []
     True
 
-    >>> Any(List.not_empty()) == [1, 2, 3]
+    >>> List.not_empty() == [1, 2, 3]
     True
-    >>> Any(List.not_empty()) == []
+    >>> List.not_empty() == []
     False
     """
 
 
 class Set(_CollectionValuesChecker, set, metaclass=_CollectionValuesCheckerMeta):
-    """Special class allowing utils.Any to check items in a set
+    """Special class enabling equality comparisons to check items in a set
 
     Examples of functionality:
 
-    >>> Any(Set.containing(1)) == {1, 2, 3}
+    >>> Set.containing(1) == {1, 2, 3}
     True
-    >>> Any(Set.containing(1)) == {4, 5, 6}
+    >>> Set.containing(1) == {4, 5, 6}
     False
 
-    >>> Any(Set.not_containing(1)) == {1, 2, 3}
+    >>> Set.not_containing(1) == {1, 2, 3}
     False
-    >>> Any(Set.not_containing(1)) == {4, 5, 6}
+    >>> Set.not_containing(1) == {4, 5, 6}
     True
 
-    >>> Any(Set.empty()) == {1, 2, 3}
+    >>> Set.empty() == {1, 2, 3}
     False
-    >>> Any(Set.empty()) == set()
+    >>> Set.empty() == set()
     True
 
-    >>> Any(Set.not_empty()) == {1, 2, 3}
+    >>> Set.not_empty() == {1, 2, 3}
     True
-    >>> Any(Set.not_empty()) == set()
+    >>> Set.not_empty() == set()
     False
+    """
+
+
+class Dict(_DictValuesChecker, dict, metaclass=_CollectionValuesCheckerMeta):
+    """Special class allowing utils.Any to check items in a dict
+
+    Examples of functionality:
+
+    >>> Dict.containing('a') == {'a': 1, 'b': 2}
+    True
+    >>> Dict.containing(a=1) == {'a': 1, 'b': 2}
+    True
+    >>> Dict.containing({'a': 1}) == {'a': 1, 'b': 2}
+    True
+    >>> Dict.containing('a') == {'b': 2}
+    False
+    >>> Dict.containing(a=1) == {'b': 2}
+    False
+    >>> Dict.containing({'a': 1}) == {'b': 2}
+    False
+
+    >>> Dict.not_containing('a') == {'a': 1, 'b': 2}
+    False
+    >>> Dict.not_containing(a=1) == {'a': 1, 'b': 2}
+    False
+    >>> Dict.not_containing({'a': 1}) == {'a': 1, 'b': 2}
+    False
+    >>> Dict.not_containing('a') == {'b': 2}
+    True
+    >>> Dict.not_containing(a=1) == {'b': 2}
+    True
+    >>> Dict.not_containing({'a': 1}) == {'b': 2}
+    True
+
+    >>> Dict.empty() == {'a': 1, 'b': 2, 'c': 3}
+    False
+    >>> Dict.empty() == {}
+    True
+
+    >>> Dict.not_empty() == {'a': 1, 'b': 2, 'c': 3}
+    True
+    >>> Dict.not_empty() == {}
+    False
+
+    """
+
+
+class Str(_CollectionValuesChecker, str, metaclass=_CollectionValuesCheckerMeta):
+    """Special class enabling equality comparisons to check items in a string
+
+    Examples of functionality:
+
+    >>> Str.containing('app') == 'apple'
+    True
+    >>> Str.containing('app') == 'happy'
+    True
+    >>> Str.containing('app') == 'banana'
+    False
+
+    >>> Str.not_containing('app') == 'apple'
+    False
+    >>> Str.not_containing('app') == 'happy'
+    False
+    >>> Str.not_containing('app') == 'banana'
+    True
+
+    >>> Str.empty() == 'hamster'
+    False
+    >>> Str.empty() == ''
+    True
+
+    >>> Str.not_empty() == 'hamster'
+    True
+    >>> Str.not_empty() == ''
+    False
+
     """