+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+### Added
+ - Allow `util.Collection` classes to be used directly in comparisons
+ - Add `util.Dict` and `util.Str` comparison classes
 
 
 ## [0.1.1] — 2020-03-30